ito na talaga yung final

+++ admin_site/views.py
@@ -644,6 +644,7 @@ def pos_receipt(request):
     pos_payment = Cart_Payment.objects.get(cart_user = request.user.role, cart_status = "not Print")
     sum_amount = Cart.objects.filter(cart_user = request.user).all().aggregate(total =Sum('cart_amount'))['total']
 
+    
     context = {
         'list_pos':pos,
         'sum_amount':sum_amount,
@@ -655,19 +656,55 @@ def pos_receipt(request):
 def pos_receipt_process(request):
     if request.method == "POST":
         get_paymentID = request.POST['get_id']
-        pos_payment = Cart_Payment.objects.get(id = get_paymentID)
-        pos_payment.cart_status = "Printed"
-        pos_payment.save()
-
         pos = Cart.objects.filter(cart_user = request.user)
-        pos.delete()
-        return redirect('admin_site:pos')
+        if Cart.objects.filter(cart_user = request.user):
+
+            for carts in pos:
+
+                products = Product.objects.get(product_code = carts.cart_pcode)
+                cart_quantity = int(carts.cart_quantity)
+                current_stock = int(products.product_stock)
+                minus_stock = cart_quantity - current_stock
+                products.product_stock = minus_stock
+                products.save()
+            
+                if products.product_stock == 0:
+                    products.product_status = "not available"
+                    products.save()
+
+           
+            pos_payment = Cart_Payment.objects.get(id = get_paymentID)
+            pos_payment.cart_status = "Printed"
+            pos_payment.save()
+
+            pos = Cart.objects.filter(cart_user = request.user)
+            pos.delete()
+            return redirect('admin_site:pos')
 
 def pos_addreceipt(request):
     if request.method == "POST":
 
         #saving to pos payment in databse
         pos_id = request.POST['get_id']
+
+        pos = Cart.objects.filter(cart_user = request.user)
+
+        for carts in pos:
+                    
+
+
+            products = Product.objects.get(product_code = carts.cart_pcode)
+                   
+            
+            if products.product_stock < carts.cart_quantity:
+                 
+                messages.error(request,('Some Products out of stock'))
+                return redirect('admin_site:pos')
+                
+            
+        
+
+
         if Cart_Payment.objects.filter(cart_user =request.user.role, cart_status="not Print"):
             messages.error(request,('receipt still not done'))
             return redirect('admin_site:pos')
@@ -682,6 +719,12 @@ def pos_addreceipt(request):
             new_Cart_Payment.cart_status = "not Print"
             new_Cart_Payment.save()
             return redirect('admin_site:pos_receipt')
+
+def pos_removeall(request):
+    pos = Cart.objects.filter(cart_user = request.user)
+    pos.delete()
+    messages.success(request,("successfully removed all"))
+    return redirect('admin_site:pos')
 
 
 def Click_receipt(request):
@@ -709,16 +752,16 @@ def minus_qty(request, productid):
         pos.cart_amount = result
         pos.save()
     
-        current_pcode = pos.cart_pcode
-        product = Product.objects.get(product_code = current_pcode)
-        current_stock = int(product.product_stock)
-        retrieve_stock = current_stock + 1
-        product.product_stock = retrieve_stock
-        product.save()
+        # current_pcode = pos.cart_pcode
+        # product = Product.objects.get(product_code = current_pcode)
+        # current_stock = int(product.product_stock)
+        # retrieve_stock = current_stock + 1
+        # product.product_stock = retrieve_stock
+        # product.save()
 
-        if product.product_stock > 0:
-                product.product_status = "available"
-                product.save()
+        # if product.product_stock > 0:
+        #         product.product_status = "available"
+        #         product.save()
         return redirect('admin_site:pos')
 
      
@@ -746,15 +789,15 @@ def add_qty(request,productid):
         pos.save()
     
 
-        product = Product.objects.get(product_code = current_pcode)
-        current_stock = int(product.product_stock)
-        minus_stock = current_stock - 1
-        product.product_stock = minus_stock
-        product.save()
+        # product = Product.objects.get(product_code = current_pcode)
+        # current_stock = int(product.product_stock)
+        # minus_stock = current_stock - 1
+        # product.product_stock = minus_stock
+        # product.save()
 
-        if product.product_stock == 0:
-            product.product_status = "not available"
-            product.save()
+        # if product.product_stock == 0:
+        #     product.product_status = "not available"
+        #     product.save()
 
         return redirect('admin_site:pos')
 
@@ -764,16 +807,16 @@ def pos_cancel(request,productid):
     if request.method == "POST":
         cancel = Cart.objects.get(id =productid)
         current_pcode = request.POST['current_pcode']
-        product = Product.objects.get(product_code = current_pcode)
+        # product = Product.objects.get(product_code = current_pcode)
 
-        current_qty = int(request.POST['current_qty'])
-        current_stock = int(product.product_stock)
+        # current_qty = int(request.POST['current_qty'])
+        # current_stock = int(product.product_stock)
 
-        return_stock = current_stock + current_qty
-        product.product_stock = return_stock
+        # return_stock = current_stock + current_qty
+        # product.product_stock = return_stock
         
 
-        product.save()
+        # product.save()
         cancel.delete()
 
         #activity log for cancelling the product
@@ -789,9 +832,9 @@ def pos_cancel(request,productid):
         pos_payment = Cart_Payment.objects.filter(cart_user =request.user.role, cart_status="not Print")
         pos_payment.delete()
 
-        if product.product_stock != 0:
-                product.product_status = "available"
-                product.save()
+        # if product.product_stock != 0:
+        #         product.product_status = "available"
+        #         product.save()
     
 
         
@@ -834,7 +877,7 @@ def cart_products(request, productid):
         current_user = request.user
         
         # minus or adding to the stock
-        diff = p_stock -  qty 
+        # diff = p_stock -  qty 
         amount_cart = p_price * qty
         reseller_cart = p_reseller_price * qty
         
@@ -866,8 +909,8 @@ def cart_products(request, productid):
         else:
 
             #updating product stcart
-            product.product_stock = diff
-            product.save()
+            # product.product_stock = diff
+            # product.save()
 
             #inserting product in pos table
             pos = Cart(cart_user=current_user, cart_pcode=pcode, cart_category= p_category,  cart_name = p_name, cart_unit= p_unit,cart_reseller_price =p_reseller_price , cart_price = p_price, cart_quantity = qty, cart_amount = amount_cart,  cart_ResellerAmount =reseller_cart )
@@ -876,9 +919,9 @@ def cart_products(request, productid):
             pos_payment = Cart_Payment.objects.filter(cart_user = request.user, cart_status="not Print")
             pos_payment.delete()
 
-            if product.product_stock == 0:
-                product.product_status = "not available"
-                product.save()
+            # if product.product_stock == 0:
+            #     product.product_status = "not available"
+            #     product.save()
